cxx-cmake: Read mappings from configuration file and rename commands
Mappings for available functionalities are read from configuration file
and stored in a dictionary along with a function to do the register.
If entry is defined and mapping is not empty then it is registered.

Additionally replace Proset* commands with CXXCMake* commands
(build, clean, run, run with args, update symbols).