cxx-cmake: Add prefix "settings" in some configuration parameters
Because of that, a lot of changes in the way they are stored internally
in the dictionary was applied. So now, cxx_cmake.properties dictionary
has three keys: "settings", "mappings" and "plugin". First two are taken
directly from configuration file and prefix of configuration parameter
means where it is stored in the dictionary. Third key "plugin" stores
values that are created using "settings" key and hardcoded values.