cxx-cmake: Remove throwing exception when parameter is empty
Some configuration parameters can not be empty.  In such cases
exception was thrown.  Throwing exception caused cxx-cmake not
to be loaded and Default Settings Object was loaded instead.
Although it can be considered good solution, I find it not in
this case. When configuration parameter can not be empty and for
some reason it is in the configuration file, then default
(hardcoded) value will be taken as if it was missing.