proset: Add Load Settings functionality
Even though Proset plugin loads project settings automatically if Vim
was started in project root path (and other conditions are met), there
was a need to be able to load project settings manually for example when
Vim was started outside of a project and we want to switch to project
without restarting Vim.

Future plan is to have Generate Project/Project Settings functionality
and generating it and then restarting Vim to load it, is rather poor
design.

ProsetRegisterSettings() now takes settings name and constructor
function name, so Settings Dictionary has to have public constructor
function which is used to create Settings Object inside Proset plugin
(internally, when needed) and not as before, where an already created
Settings Object was transferred to that function. This has an advantage:
Settings Object is created only when needed and only one, not all of
them.

Add new functions:

- ProsetLoadSettings(),
- ProsetReloadSettings(),
- ProsetCloseSettings().

Changes:

- ProsetGetProperties() and ProsetGetConfiguration() returns deepcopy.

When two Settings Objects are registered with the same name, disable
Proset plugin by setting s:storage to {}.

When current directory is going to be changed, close the project
settings.

If you use plugin incorrectly or something not dependent on you
happened, you will get one of possible errors in the form of a message
"proset-EN: ...", where N is a number.

In case of success or just information you will get it in the form of a
message: "proset: ...".

Gloabl variables g:proset_directory, g:proset_file are locked, so you
can not change values of the variables from Vim command line mode. Set
them in vimrc if you want to have values other than defaults.  Both
variables are validated, if any of those is incorrect you will get an
error message.

Documentations of Proset and cxx-cmake were updated as well.