cxx-cmake: Remove mappings "normal" subsection
Removed "mappings.normal" section and leave all the mappings directly in
the "mappings" section. If for some reason any module's mapping will
need to have also mapping for visual or other Vim mode, then it will use
either mapping for normal mode, or new one telling that it is for that
mode will be added.