cxx-cmake: Make internal storage of "properties.settings" section oriented
Before that change, there were:

- s:cxx_cmake.properties.settings - "flat" structure of all settings,
- s:cxx_cmake.properties.mappings - "flat" structure of all mappings,
- s:cxx_cmake.properties.internal.

After this change, there are

- s:cxx_cmake.properties.configuration - "section oriented" structure of
  settings and mappings,
- s:cxx_cmake.properties.internal.

So, "internal" was not changed, "mappings" was removed and "settings"
was renamed to "configuration" and redesigned to be very similar to JSON
configuration file format (yet, there is more information), so for example:

Before:

- s:cxx_cmake.properties.settings.build_directory
- s:cxx_cmake.properties.mappings.build

After:

- s:cxx_cmake.properties.configuration.build.settings.build_directory
- s:cxx_cmake.properties.configuration.build.mappings.build

Even though in current solution to get settings or mappings you need to
go through more levels of data structures, it is "section oriented"
which means "properties.configuration" stores sections like: "build",
"run", "ctags", etc, and all the configuration of such sections is
there, not in "flat" structures (with big possibility of name clash).
Extending (adding new sections, modifying already existing) seems to be
much easier now.

So now, each section (if it is not top-level entry) should be defined
like this:

    {"settings": {}, "mappings": {}}

In "settings" you can store whatever you want, but in "mappings" there
is strict format how it has to be implemented due to adding and removing
mappings automatically in enable() and disable(). See the code for
details. Any of them may be empty.