cxx-cmake: Store sections in object oriented fashion
Move most cxx-cmake functions to dedicated autoload functions.  Most
functions of cxx-cmake Settings Dictionary were stored in cxx-cmake.vim
file. It turned out that extending it caused this file to get bigger and
bigger and having all the stuff from all the modules in one file was not
so good.

Now, all the functions corresponding to concrete module are stored in
its own autoload file with name of the module, for example:

 autoload/proset/settings/cxx_cmake/run.vim

All the configuration of "run" section is placed there.

All of these files follow the same structure scheme, to be easily
navigated and because of the nature of this object oriented approach.

So, "run" section module (and others as well) object has:

- public constructor (autoload scheme)  - create module object
- run.enable()                          - enable module
- run.disable()                         - disable module (undo enable)
- run.get_configuration()               - configuration of the module

In cxx-cmake.vim there is creation of "run" object using public
constructor, calling run.enable(), run.disable() and run.get_configuration().

Additionally cxx-cmake object is extended with "modules" dictionary,
where all section objects are stored.

Some functions from cxx-cmake.vim were moved to util file because they
are used across modules files.

For now, there are following modules/sections:

- alternate_file.vim
- build.vim
- create_header.vim
- create_header_source.vim
- create_source.vim
- cscope.vim
- ctags.vim
- run.vim
- source.vim
- symbols.vim