cxx-cmake: Do not add or remove mapping if it is empty
Mapping sequence is trimmed first and then checked if it is empty.
If so, do not add or remove such mapping.  When settings.json file
did not have "cscope.mappings" section defined, which means mappings
were disabled for cscope, mappings for :cs command were set anyway.
This caused delay when switching Vim to command mode.