cxx-cmake: Reuse run output window and delete old buffer (#2)
When CXXCMakeRun (either with or without args) was called, new window
was created and filled with process output (term_start()). In some
scenarios this behaviour is good or even preferred, yet for most of the
time it is quite interruptive.

If there is no output window, one will be created. If there is already
one that was created by term_start() then this window will be used for
new process execution. Old buffer is deleted.

closes: #2